Added a CompanyFactory test for missing registrationDate
